% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knutils.R
\name{computeprob}
\alias{computeprob}
\title{Computes the probability that a model parameter is <=k from inla marginals}
\usage{
computeprob(func, k)
}
\arguments{
\item{func}{is the inla marginals of the model parameter}

\item{k}{is the cutoff}
}
\value{
probability model coefficient <=k
}
\description{
This function will be used to calculate the P(coeficient variable cluster <=0)
}
