% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions2.R
\name{CreateGridDClusterm}
\alias{CreateGridDClusterm}
\title{Creates grid over the study area.}
\usage{
CreateGridDClusterm(stfdf, radius, step)
}
\arguments{
\item{stfdf}{spatio-temporal class object containing the data.}

\item{radius}{maximum radius of the clusters.}

\item{step}{step of the grid.}
}
\value{
two columns matrix where each row represents a point of the grid.
}
\description{
If the argument thegrid of DetectClustersModel() is null, this function is
used to create a rectangular grid with a given step.
If step is NULL the step used is equal to 0.2*radius.
The grid contains the coordinates of the centers of the clusters explored.
}
