% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knutils.R
\name{get.stclusters}
\alias{get.stclusters}
\title{Gets areas in a spatio-temporal cluster}
\usage{
get.stclusters(stfdf, results)
}
\arguments{
\item{stfdf}{A sp or spacetime object with the information about the data.}

\item{results}{Results from a call to DetectClusterModel}
}
\value{
A list with as many elements as clusters in 'results'
}
\description{
This function is similar to get.knclusters but it also allows
for spatio-temporal clusters.
}
