\name{knbinary}
\alias{knbinary}
\title{This function makes binary variables from the clusters...}
\usage{knbinary(d, knresults)}
\description{This function makes binary variables from the clusters}
\value{whatever the return is}
\arguments{\item{d}{whatever d is}
\item{knresults}{whatever knresults is}}
