% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knutils.R
\name{PlotClustersNoOverlap}
\alias{PlotClustersNoOverlap}
\title{Plots the clusters that do not overlap.}
\usage{
PlotClustersNoOverlap(statsAllClustersNoOverlap, colors, map)
}
\arguments{
\item{statsAllClustersNoOverlap}{data frame with information of the detected clusters
that no overlap. Each row represents the information of one of the clusters.
It contains the coordinates of the center, the size, the start
and end dates, the log-likelihood ratio, a boolean indicating if it is a
cluster (TRUE in all cases), and the p-value of the cluster.}

\item{colors}{vector with the colors of the clusters.}

\item{map}{SpatialPolygonsDataFrame with the polygons of the map.}
}
\value{
plots of the detected clusters for each start date.
}
\description{
This function plots the detected clusters that do not overlap.
There are as many windows as different start dates. All clusters
with the same start date are represented in the same window.
}
