\name{slimknclusters}
\alias{slimknclusters}
\title{This function slims the number of clusters down.}
\usage{slimknclusters(d, knresults, minsize=1)}
\description{This function slims the number of clusters down.}
\details{The spatial scan statistic is known to detect duplicated
clusters. This function aims to reduce the number of clusters
by removing duplicated and overlapping clusters.

The main criteria to choose the 'primary' clusters 
is the likelihood ratio}
\value{whatever this does}
\arguments{\item{d}{whatever d is}
\item{knresults}{whatever knresults is}
\item{minsize}{whateverthe minsize}}
