\name{get.stclusters}
\alias{get.stclusters}
\title{Gets areas in a spatio-temporal cluster...}
\usage{get.stclusters(stfdf, results)}
\description{Gets areas in a spatio-temporal cluster
This function is similar to get.knclusters but it also allows
for spatio-temporal clusters.}
\value{A list with as many elements as clusters in 'results'}
\arguments{\item{stfdf}{A sp or spacetime object with the information about the data.}
\item{results}{Results from a call to DetectClusterModel}}
