\name{mergeknclusters}
\alias{mergeknclusters}
\title{Merge clusters so that they are identifed as factors.}
\usage{mergeknclusters(d, knresults)}
\description{Merge clusters so that they are identifed as factors.
Clusters should not overlap.}
\value{whatever the return is}
\arguments{\item{d}{whatever d is}
\item{knresults}{whatever knresults is}}
