% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions1.R
\name{SelectStatsAllClustersNoOverlap}
\alias{SelectStatsAllClustersNoOverlap}
\title{Removes the overlapping clusters.}
\usage{
SelectStatsAllClustersNoOverlap(stfdf, statsAllClusters)
}
\arguments{
\item{stfdf}{spatio-temporal class object containing the data.}

\item{statsAllClusters}{data frame with information of the detected
clusters obtained with DetectClustersModel().}
}
\value{
data frame with the same information than statsAllClusters but only
for clusters that do not overlap.
}
\description{
Function DetectClustersModel() detects duplicated clusters.
This function reduces the number of clusters by removing the overlapping
clusters.
}
\examples{
library("DClusterm")
data("brainNM")
data("brainNM_clusters")

SelectStatsAllClustersNoOverlap(brainst, nm.cl1)
}
