% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knutils.R
\name{slimknclusters}
\alias{slimknclusters}
\title{Remove overlapping clusters}
\usage{
slimknclusters(d, knresults, minsize = 1)
}
\arguments{
\item{d}{Data.frame with data used in the detection of clusters.}

\item{knresults}{Object returned by function opgam() with the clusters detected.}

\item{minsize}{Minimum size of cluster (default to 1).}
}
\value{
A subset of knresults with non-overlaping clusters of at least
minsize size.
}
\description{
This function slims the number of clusters down.
The spatial scan statistic is known to detect duplicated
clusters. This function aims to reduce the number of clusters
by removing duplicated and overlapping clusters.
}
